feat: int로 받을 수 있는 타입을 Integer로 변경

+++ src/main/java/com/example/feedprep/domain/feedbackreview/dto/FeedbackReviewRequestDto.java
@@ -14,7 +14,7 @@ public class FeedbackReviewRequestDto {
 
 	@Min(value = 0L)
 	@Max(value = 5L)
-	private Long rating;
+	private Integer rating;
 
 	private String content;
 }
